feature008: Added production in process service
