+++ formulapanadera/src/app/core/core.module.ts
@@ -23,6 +23,7 @@ import { ProductionService } from "./services/production.service";
 import { ProductionStorageService } from "./services/storage/production.service";
 import { TimeService } from "./services/time.service";
 import { ProductionInProcessStorageService } from "./services/storage/production-in-process.service";
+import { ProductionInProcessService } from "./services/production-in-process.service";
 
 @NgModule({
   imports: [
@@ -44,6 +45,7 @@ import { ProductionInProcessStorageService } from "./services/storage/production
     IngredientService,
     IngredientCRUDService,
     ProductionService,
+    ProductionInProcessService,
     ProductionCRUDService,
     ProductionStorageService,
     ProductionInProcessStorageService,
